\name{nncRda-package}
\alias{nncRda-package}
\docType{package}

\title{Improved RDA Using nnc With Cross-Validation}

\description{
Improved RDA (Regularized discriminant analysis) for class prediction 
in microarrays using nearest neighbour covariates.
A function for simulating microarray experiments is provided. 
}

\details{
\tabular{ll}{
Package: \tab nncRda\cr
Type: \tab Package\cr
Version: \tab 1.87\cr
Date: \tab 2010-10-04\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}

\author{
M. S. Islam and A. I. McLeod
 <aimcleod@uwo.ca>
}

\references{
Islam, M. S. (2008). Periodicity, Change Detection and Prediction in Microarrays.
Ph.D. Thesis, The University of Western Ontario.
}

\seealso{
\code{\link{FitRDA}}, 
\code{\link{nncRda}}
}

\examples{
\dontrun{
2+2
}
} 

\keyword{ package }


